fix: add toggle with icon checkmark and cross
