+++ app/src/components/HelpView.tsx
@@ -14,7 +14,7 @@ const HelpView = ({ title, description, link }: Props) => {
   return (
     <View className="flex-1 bg-white p-4 pb-16">
       <Text className={mergeClassNames(typography.textXlBold, "mb-4 text-cnam-primary-950")}>{title}</Text>
-      <Text className={mergeClassNames(typography.textLgRegular, "text-cnam-primary-950")}>{description}</Text>
+      <Text className={mergeClassNames(typography.textMdMedium, "text-cnam-primary-800")}>{description}</Text>
       {link && (
         <TouchableOpacity
           onPress={() => {
